ref(server): Simplify HTTP server and service startup
