+++ relay-server/src/actors/server.rs
@@ -1,20 +1,148 @@
+use actix::Recipient;
+use actix_web::server::{self, StopServer};
+use actix_web::App;
+use anyhow::{Context, Result};
+use futures01::Future;
+use listenfd::ListenFd;
+
 use relay_config::Config;
 use relay_statsd::metric;
-use relay_system::{Controller, Service, Shutdown};
+use relay_system::{Addr, Controller, Service, Shutdown};
 
-use crate::service::HttpServer;
+use crate::middlewares::{
+    AddCommonHeaders, ErrorHandlers, Metrics, ReadRequestMiddleware, SentryMiddleware,
+};
+use crate::service::ServiceState;
 use crate::statsd::RelayCounters;
 
+// TODO(ja): Split this into Server and Service error.
+use crate::service::ServerError;
+
+/// The actix app type for the relay web service.
+pub type ServiceApp = App<ServiceState>;
+
+fn make_app(state: ServiceState) -> ServiceApp {
+    App::with_state(state)
+        .middleware(SentryMiddleware::new())
+        .middleware(Metrics)
+        .middleware(AddCommonHeaders)
+        .middleware(ErrorHandlers)
+        .middleware(ReadRequestMiddleware)
+        .configure(crate::endpoints::configure_app)
+}
+
+fn dump_listen_infos<H, F>(server: &server::HttpServer<H, F>)
+where
+    H: server::IntoHttpHandler + 'static,
+    F: Fn() -> H + Send + Clone + 'static,
+{
+    relay_log::info!("spawning http server");
+    for (addr, scheme) in server.addrs_with_scheme() {
+        relay_log::info!("  listening on: {}://{}/", scheme, addr);
+    }
+}
+
+fn listen<H, F>(
+    server: server::HttpServer<H, F>,
+    config: &Config,
+) -> Result<server::HttpServer<H, F>>
+where
+    H: server::IntoHttpHandler + 'static,
+    F: Fn() -> H + Send + Clone + 'static,
+{
+    Ok(
+        match ListenFd::from_env()
+            .take_tcp_listener(0)
+            .context(ServerError::ListenFailed)?
+        {
+            Some(listener) => server.listen(listener),
+            None => server
+                .bind(config.listen_addr())
+                .context(ServerError::BindFailed)?,
+        },
+    )
+}
+
+#[cfg(feature = "ssl")]
+fn listen_ssl<H, F>(
+    mut server: server::HttpServer<H, F>,
+    config: &Config,
+) -> Result<server::HttpServer<H, F>>
+where
+    H: server::IntoHttpHandler + 'static,
+    F: Fn() -> H + Send + Clone + 'static,
+{
+    if let (Some(addr), Some(path), Some(password)) = (
+        config.tls_listen_addr(),
+        config.tls_identity_path(),
+        config.tls_identity_password(),
+    ) {
+        use native_tls::{Identity, TlsAcceptor};
+        use std::fs::File;
+        use std::io::Read;
+
+        let mut file = File::open(path).unwrap();
+        let mut data = vec![];
+        file.read_to_end(&mut data).unwrap();
+        let identity =
+            Identity::from_pkcs12(&data, password).context(ServerError::TlsInitFailed)?;
+
+        let acceptor = TlsAcceptor::builder(identity)
+            .build()
+            .context(ServerError::TlsInitFailed)?;
+
+        server = server
+            .bind_tls(addr, acceptor)
+            .context(ServerError::BindFailed)?;
+    }
+
+    Ok(server)
+}
+
+#[cfg(not(feature = "ssl"))]
+fn listen_ssl<H, F>(
+    server: server::HttpServer<H, F>,
+    config: &Config,
+) -> Result<server::HttpServer<H, F>, ServerError>
+where
+    H: server::IntoHttpHandler + 'static,
+    F: Fn() -> H + Send + Clone + 'static,
+{
+    if config.tls_listen_addr().is_some()
+        || config.tls_identity_path().is_some()
+        || config.tls_identity_password().is_some()
+    {
+        Err(ServerError::TlsNotSupported.into())
+    } else {
+        Ok(server)
+    }
+}
+
+// TODO(ja): Document
 pub struct ServerService {
-    http_server: HttpServer,
+    http_server: Recipient<StopServer>,
 }
 
 impl ServerService {
-    pub fn start(config: Config) -> anyhow::Result<()> {
+    pub fn start(config: &Config, service: ServiceState) -> anyhow::Result<Addr<()>> {
         metric!(counter(RelayCounters::ServerStarting) += 1);
-        let http_server = HttpServer::start(config)?;
-        Self { http_server }.start();
-        Ok(())
+
+        let mut server = server::new(move || make_app(service.clone()));
+        server = server
+            .workers(config.cpu_concurrency())
+            .shutdown_timeout(config.shutdown_timeout().as_secs() as u16)
+            .keep_alive(config.keepalive_timeout().as_secs() as usize)
+            .maxconn(config.max_connections())
+            .maxconnrate(config.max_connection_rate())
+            .backlog(config.max_pending_connections())
+            .disable_signals();
+
+        server = listen(server, config)?;
+        server = listen_ssl(server, config)?;
+        dump_listen_infos(&server);
+
+        let http_server = server.start().recipient();
+        Ok(Self { http_server }.start())
     }
 }
 
@@ -24,13 +152,17 @@ impl Service for ServerService {
     fn spawn_handler(self, _rx: relay_system::Receiver<Self::Interface>) {
         tokio::spawn(async move {
             let mut shutdown = Controller::shutdown_handle();
+
             loop {
-                tokio::select! {
-                    Shutdown { timeout } = shutdown.notified() => {
-                       self.http_server.shutdown(timeout.is_some());
-                    },
-                    else => break,
-                }
+                let Shutdown { timeout } = shutdown.notified().await;
+                let graceful = timeout.is_some();
+
+                // We assume graceful shutdown if we're given a timeout. The actix-web http server is
+                // configured with the same timeout, so it will match. Unfortunately, we have to drop any
+                // errors  and replace them with the generic `TimeoutError`.
+                relay_log::info!("Shutting down HTTP server"); // TODO(ja): Where was this coming from?
+                self.http_server.send(StopServer { graceful }).wait().ok();
+                // TODO(ja): harmful. wait() blocks the main runtime.
             }
         });
     }
