+++ relay-server/src/service.rs
@@ -1,12 +1,7 @@
 use std::fmt;
 use std::sync::Arc;
 
-use actix::Recipient;
-use actix_web::server::StopServer;
-use actix_web::{server, App};
 use anyhow::{Context, Result};
-use futures01::Future;
-use listenfd::ListenFd;
 use once_cell::race::OnceBox;
 use tokio::runtime::Runtime;
 
@@ -14,7 +9,7 @@ use relay_aws_extension::AwsExtension;
 use relay_config::Config;
 use relay_metrics::{Aggregator, AggregatorService};
 use relay_redis::RedisPool;
-use relay_system::{Addr, Configure, Controller, Service};
+use relay_system::{Addr, Service};
 
 use crate::actors::envelopes::{EnvelopeManager, EnvelopeManagerService};
 use crate::actors::health_check::{HealthCheck, HealthCheckService};
@@ -27,12 +22,12 @@ use crate::actors::relays::{RelayCache, RelayCacheService};
 use crate::actors::store::StoreService;
 use crate::actors::test_store::{TestStore, TestStoreService};
 use crate::actors::upstream::{UpstreamRelay, UpstreamRelayService};
-use crate::middlewares::{
-    AddCommonHeaders, ErrorHandlers, Metrics, ReadRequestMiddleware, SentryMiddleware,
-};
 use crate::utils::BufferGuard;
 
 pub static REGISTRY: OnceBox<Registry> = OnceBox::new();
+
+// TODO(ja): Remove this?
+pub use crate::actors::server::ServiceApp;
 
 /// Indicates the type of failure of the server.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, thiserror::Error)]
@@ -221,154 +216,5 @@ impl ServiceState {
     /// buffer. See [`BufferGuard`] for more information.
     pub fn buffer_guard(&self) -> &BufferGuard {
         &self.buffer_guard
-    }
-}
-
-/// The actix app type for the relay web service.
-pub type ServiceApp = App<ServiceState>;
-
-fn make_app(state: ServiceState) -> ServiceApp {
-    App::with_state(state)
-        .middleware(SentryMiddleware::new())
-        .middleware(Metrics)
-        .middleware(AddCommonHeaders)
-        .middleware(ErrorHandlers)
-        .middleware(ReadRequestMiddleware)
-        .configure(crate::endpoints::configure_app)
-}
-
-fn dump_listen_infos<H, F>(server: &server::HttpServer<H, F>)
-where
-    H: server::IntoHttpHandler + 'static,
-    F: Fn() -> H + Send + Clone + 'static,
-{
-    relay_log::info!("spawning http server");
-    for (addr, scheme) in server.addrs_with_scheme() {
-        relay_log::info!("  listening on: {}://{}/", scheme, addr);
-    }
-}
-
-fn listen<H, F>(
-    server: server::HttpServer<H, F>,
-    config: &Config,
-) -> Result<server::HttpServer<H, F>>
-where
-    H: server::IntoHttpHandler + 'static,
-    F: Fn() -> H + Send + Clone + 'static,
-{
-    Ok(
-        match ListenFd::from_env()
-            .take_tcp_listener(0)
-            .context(ServerError::ListenFailed)?
-        {
-            Some(listener) => server.listen(listener),
-            None => server
-                .bind(config.listen_addr())
-                .context(ServerError::BindFailed)?,
-        },
-    )
-}
-
-#[cfg(feature = "ssl")]
-fn listen_ssl<H, F>(
-    mut server: server::HttpServer<H, F>,
-    config: &Config,
-) -> Result<server::HttpServer<H, F>>
-where
-    H: server::IntoHttpHandler + 'static,
-    F: Fn() -> H + Send + Clone + 'static,
-{
-    if let (Some(addr), Some(path), Some(password)) = (
-        config.tls_listen_addr(),
-        config.tls_identity_path(),
-        config.tls_identity_password(),
-    ) {
-        use native_tls::{Identity, TlsAcceptor};
-        use std::fs::File;
-        use std::io::Read;
-
-        let mut file = File::open(path).unwrap();
-        let mut data = vec![];
-        file.read_to_end(&mut data).unwrap();
-        let identity =
-            Identity::from_pkcs12(&data, password).context(ServerError::TlsInitFailed)?;
-
-        let acceptor = TlsAcceptor::builder(identity)
-            .build()
-            .context(ServerError::TlsInitFailed)?;
-
-        server = server
-            .bind_tls(addr, acceptor)
-            .context(ServerError::BindFailed)?;
-    }
-
-    Ok(server)
-}
-
-#[cfg(not(feature = "ssl"))]
-fn listen_ssl<H, F>(
-    server: server::HttpServer<H, F>,
-    config: &Config,
-) -> Result<server::HttpServer<H, F>, ServerError>
-where
-    H: server::IntoHttpHandler + 'static,
-    F: Fn() -> H + Send + Clone + 'static,
-{
-    if config.tls_listen_addr().is_some()
-        || config.tls_identity_path().is_some()
-        || config.tls_identity_password().is_some()
-    {
-        Err(ServerError::TlsNotSupported.into())
-    } else {
-        Ok(server)
-    }
-}
-
-/// Keeps the address to the running http servers and helps with start/stop handling.
-pub struct HttpServer(Recipient<StopServer>);
-
-impl HttpServer {
-    /// Given a relay config spawns the server together with all actors and lets them run forever.
-    ///
-    /// Effectively this boots the server.
-    pub fn start(config: Config) -> Result<Self> {
-        let config = Arc::new(config);
-
-        Controller::from_registry().do_send(Configure {
-            shutdown_timeout: config.shutdown_timeout(),
-        });
-
-        let state = ServiceState::start(config.clone())?;
-        let mut server = server::new(move || make_app(state.clone()));
-        server = server
-            .workers(config.cpu_concurrency())
-            .shutdown_timeout(config.shutdown_timeout().as_secs() as u16)
-            .keep_alive(config.keepalive_timeout().as_secs() as usize)
-            .maxconn(config.max_connections())
-            .maxconnrate(config.max_connection_rate())
-            .backlog(config.max_pending_connections())
-            .disable_signals();
-
-        server = listen(server, &config)?;
-        server = listen_ssl(server, &config)?;
-
-        dump_listen_infos(&server);
-        let recipient = server.start().recipient();
-        Ok(Self(recipient))
-    }
-
-    /// Triggers the shutdown process by sending [`actix_web::server::StopServer`] to the running http server.
-    pub fn shutdown(&self, graceful: bool) {
-        let Self(recipient) = self;
-        relay_log::info!("Shutting down HTTP server");
-        recipient.send(StopServer { graceful }).wait().ok();
-    }
-}
-
-impl fmt::Debug for HttpServer {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_tuple("HttpServer")
-            .field(&"actix::Recipient<actix_web::server::StopServer>")
-            .finish()
     }
 }
